refactor: rename ru to ru_RU
