+++ packages/s2-core/src/common/i18n/index.ts
@@ -1,11 +1,11 @@
 import { get, merge } from 'lodash';
-import { ZH_CN as BASE_ZH_CN } from './zh_CN';
 import { EN_US as BASE_EN_US } from './en_US';
-import { RU as BASE_RU } from './ru';
+import { RU as BASE_RU } from './ru_RU';
+import { ZH_CN as BASE_ZH_CN } from './zh_CN';
 
 const DEFAULT_LANG: LangType = 'zh_CN';
 
-export type LangType = 'zh_CN' | 'en_US' | 'ru';
+export type LangType = 'zh_CN' | 'en_US' | 'ru_RU';
 
 export type LocaleType = {
   [K in LangType]: Record<string, string>;
@@ -16,7 +16,7 @@ let lang: LangType = DEFAULT_LANG;
 let locale: LocaleType = {
   zh_CN: BASE_ZH_CN,
   en_US: BASE_EN_US,
-  ru: BASE_RU,
+  ru_RU: BASE_RU,
 };
 
 export const getLang = () => lang;
