+++ packages/s2-shared/src/constant/i18n/index.ts
@@ -1,10 +1,10 @@
 import type { LocaleType } from '@antv/s2';
 import { EN_US } from './en_US';
+import { RU } from './ru_RU';
 import { ZH_CN } from './zh_CN';
-import { RU } from './ru';
 
 export const Locale: LocaleType = {
   zh_CN: ZH_CN,
   en_US: EN_US,
-  ru: RU,
+  ru_RU: RU,
 };
