+++ packages/s2-shared/src/constant/i18n/ru_RU.ts
