+++ s2-site/docs/manual/basic/i18n.en.md
@@ -19,7 +19,7 @@ import { setLang, PivotSheet } from '@antv/s2'
 
 setLang('en_US')
 // setLang('zh_CN')
-// setLang('ru');
+// setLang('ru_RU');
 
 const s2 = new PivotSheet()
 s2.render()
@@ -39,7 +39,7 @@ const locale = {
   en_US: {
     test: 'test',
   },
-  ru: {
+  ru_RU: {
     test: 'тест',
   },
 };
