+++ s2-site/docs/manual/basic/i18n.zh.md
@@ -14,6 +14,7 @@ import { setLang, PivotSheet } from '@antv/s2'
 
 setLang('en_US')
 // setLang('zh_CN')
+// setLang('ru_RU');
 
 const s2 = new PivotSheet()
 s2.render()
@@ -32,6 +33,9 @@ const locale = {
   },
   en_US: {
     test: 'test',
+  },
+  ru_RU: {
+    test: 'тест',
   },
 };
